% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step4_MetaboAnnotation.R
\name{annota_Data_Tandem}
\alias{annota_Data_Tandem}
\title{Title annota_Data_Tandem}
\usage{
annota_Data_Tandem(AnnotaParamTandem)
}
\arguments{
\item{AnnotaParamTandem}{The parameters of metabolite annotation of MS/MS.}
}
\value{
The table of metabolite annotation of MS/MS.
}
\description{
Title annota_Data_Tandem
}
\examples{
ParentMass <- AnnotaData$ParentMass
TandemData <- AnnotaData$TandemData
AnnotaParamTandem <- Annota_Tandem(ParentMass, TandemData)
annota_Data_Tandem(AnnotaParamTandem)
}
