% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step4_MetaboAnnotation.R
\name{Annota_Tandem}
\alias{Annota_Tandem}
\title{Title Annota_Tandem}
\usage{
Annota_Tandem(
  ParentMass = 181.04,
  TandemData,
  massTandem = 0.1,
  toleUnitTandem = 1,
  massmzTandem = 0.5,
  toleUnitmzTandem = 1,
  ModeTandem = "Positive",
  ionEnergy = "low(10V)"
)
}
\arguments{
\item{ParentMass}{The m/z value of parent ion for metabolite annotation of MS/MS.}

\item{TandemData}{The MS/MS peak list (including the matrix of m/z values and intensities of MS/MS) for metabolite annotation of MS/MS.}

\item{massTandem}{The tolerance of parent ion for metabolite annotation of MS/MS, and the unit is ppm or Da.}

\item{toleUnitTandem}{The unit of parent ion for metabolite annotation of MS/MS. The unit is 1 for Da and 2 for ppm.}

\item{massmzTandem}{The tolerance of MS/MS peak for metabolite annotation of MS/MS, and the unit is ppm or Da.}

\item{toleUnitmzTandem}{The unit of MS/MS peak for metabolite annotation of MS/MS. The unit is 1 for Da and 2 for ppm.}

\item{ModeTandem}{The ionization mode for metabolite annotation of tandem mass spectrum. The mode can be Positive or Negative.}

\item{ionEnergy}{The CID energy for metabolite annotation of MS/MS. The mode can be low(10V), Medium(25V), High(40V), or All.}
}
\value{
The parameters for metabolite annotation of MS/MS.
}
\description{
Title Annota_Tandem
}
\examples{
ParentMass <- AnnotaData$ParentMass
TandemData <- AnnotaData$TandemData
AnnotaParamTandem <- Annota_Tandem(ParentMass, TandemData)
}
