% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step4_MetaboAnnotation.R
\name{Annota_Tandem_plot}
\alias{Annota_Tandem_plot}
\title{Title Annota_Tandem_plot}
\usage{
Annota_Tandem_plot(AnnotaParamTandem, TandemData)
}
\arguments{
\item{AnnotaParamTandem}{The parameters of metabolite annotation of MS/MS.}

\item{TandemData}{The MS/MS peak list (including the matrix of m/z values and intensities of MS/MS) for metabolite annotation of MS/MS.}
}
\value{
The plot of results for metabolite annotation of MS/MS.
}
\description{
Title Annota_Tandem_plot
}
\examples{
ParentMass <- AnnotaData$ParentMass
TandemData <- AnnotaData$TandemData
AnnotaParamTandem <- Annota_Tandem(ParentMass, TandemData)
Annota_Tandem_plot(AnnotaParamTandem, TandemData)
}
