% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step4_MetaboAnnotation.R
\name{annotaDataTandem}
\alias{annotaDataTandem}
\title{Title annotaDataTandem}
\usage{
annotaDataTandem(AnnotaParamTandem)
}
\arguments{
\item{AnnotaParamTandem}{The parameters of metabolite annotation of Tandem mass spectrum.}
}
\value{
The table of results for metabolite annotation of Tandem mass spectrum.
}
\description{
Title annotaDataTandem
}
\examples{
Parent_mass <- AnnotaData$Parent_mass
TandemData <- AnnotaData$TandemData
AnnotaParamTandem <- AnnotaTandem(Parent_mass, TandemData)
annotaDataTandem(AnnotaParamTandem)
}
