% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step4_MetaboAnnotation.R
\name{annotaDataTandom}
\alias{annotaDataTandom}
\title{Title annotaDataTandom}
\usage{
annotaDataTandom(AnnotaParamTandom)
}
\arguments{
\item{AnnotaParamTandom}{The parameters of metabolite annotation of tandom mass spectrum.}
}
\value{
The table of results for metabolite annotation of tandom mass spectrum.
}
\description{
Title annotaDataTandom
}
\examples{
Parent_mass <- AnnotaData$Parent_mass
TandomData <- AnnotaData$TandomData
AnnotaParamTandom <- AnnotaTandom(Parent_mass, TandomData)
annotaDataTandom(AnnotaParamTandom)
}
