% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step4_MetaboAnnotation.R
\name{AnnotaTandem}
\alias{AnnotaTandem}
\title{Title AnnotaTandem}
\usage{
AnnotaTandem(
  Parent_mass = 181.04,
  TandemData,
  massTandem = 0.1,
  toleUnitTandem = 1,
  massmzTandem = 0.5,
  toleUnitmzTandem = 1,
  ModeTandem = "Positive",
  ionEnergy = "low(10V)"
)
}
\arguments{
\item{Parent_mass}{The value of parent ion mass for metabolite annotation of Tandem mass spectrum.}

\item{TandemData}{The value of MS/MS peak list (m/z & Intensity) for metabolite annotation of Tandem mass spectrum.}

\item{massTandem}{The tolerance of parent ion mass for metabolite annotation of Tandem mass spectrum.}

\item{toleUnitTandem}{The unit of parent ion mass for metabolite annotation of Tandem mass spectrum. The unit could be 1 for Da, or 2 for ppm.}

\item{massmzTandem}{The tolerance of MS/MS peak mass for metabolite annotation of Tandem mass spectrum.}

\item{toleUnitmzTandem}{The unit of MS/MS peak mass for metabolite annotation of Tandem mass spectrum. The unit could be 1 for Da, or 2 for ppm.}

\item{ModeTandem}{The ionization mode for metabolite annotation of Tandem mass spectrum. The mode could be Positive or Negative.}

\item{ionEnergy}{The CID energy for metabolite annotation of Tandem mass spectrum. The mode could be low(10V) Medium(25V) High(40V) or All.}
}
\value{
The parameters of metabolite annotation of Tandem mass spectrum.
}
\description{
Title AnnotaTandem
}
\examples{
Parent_mass <- AnnotaData$Parent_mass
TandemData <- AnnotaData$TandemData
AnnotaTandem(Parent_mass, TandemData)
}
