% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step4_MetaboAnnotation.R
\name{AnnotaTandem_plot}
\alias{AnnotaTandem_plot}
\title{Title AnnotaTandem_plot}
\usage{
AnnotaTandem_plot(AnnotaParamTandem, TandemData)
}
\arguments{
\item{AnnotaParamTandem}{The parameters of metabolite annotation of Tandem mass spectrum.}

\item{TandemData}{The value of MS/MS peak list (m/z & Intensity) for metabolite annotation of Tandem mass spectrum.}
}
\value{
The plot of results for metabolite annotation of Tandem mass spectrum.
}
\description{
Title AnnotaTandem_plot
}
\examples{
Parent_mass <- AnnotaData$Parent_mass
TandemData <- AnnotaData$TandemData
AnnotaParamTandem <- AnnotaTandem(Parent_mass, TandemData)
annotaDataTandem(AnnotaParamTandem)
AnnotaTandem_plot(AnnotaParamTandem, TandemData)
}
