% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step4_MetaboAnnotation.R
\name{AnnotaTandom}
\alias{AnnotaTandom}
\title{Title AnnotaTandom}
\usage{
AnnotaTandom(
  Parent_mass = 181.04,
  TandomData,
  massTandom = 0.1,
  toleUnitTandom = 1,
  massmzTandom = 0.5,
  toleUnitmzTandom = 1,
  ModeTandom = "Positive",
  ionEnergy = "low(10V)"
)
}
\arguments{
\item{Parent_mass}{The value of parent ion mass for metabolite annotation of tandom mass spectrum.}

\item{TandomData}{The value of MS/MS peak list (m/z & Intensity) for metabolite annotation of tandom mass spectrum.}

\item{massTandom}{The tolerance of parent ion mass for metabolite annotation of tandom mass spectrum.}

\item{toleUnitTandom}{The unit of parent ion mass for metabolite annotation of tandom mass spectrum. The unit could be 1 for Da, or 2 for ppm.}

\item{massmzTandom}{The tolerance of MS/MS peak mass for metabolite annotation of tandom mass spectrum.}

\item{toleUnitmzTandom}{The unit of MS/MS peak mass for metabolite annotation of tandom mass spectrum. The unit could be 1 for Da, or 2 for ppm.}

\item{ModeTandom}{The ionization mode for metabolite annotation of tandom mass spectrum. The mode could be Positive or Negative.}

\item{ionEnergy}{The CID energy for metabolite annotation of tandom mass spectrum. The mode could be low(10V) Medium(25V) High(40V) or All.}
}
\value{
The parameters of metabolite annotation of tandom mass spectrum.
}
\description{
Title AnnotaTandom
}
\examples{
Parent_mass <- AnnotaData$Parent_mass
TandomData <- AnnotaData$TandomData
AnnotaTandom(Parent_mass, TandomData)
}
