% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step4_MetaboAnnotation.R
\name{annotaTandom_plot}
\alias{annotaTandom_plot}
\title{Title annotaTandom_plot}
\usage{
annotaTandom_plot(AnnotaParamTandom, TandomData)
}
\arguments{
\item{AnnotaParamTandom}{The parameters of metabolite annotation of tandom mass spectrum.}

\item{TandomData}{The value of MS/MS peak list (m/z & Intensity) for metabolite annotation of tandom mass spectrum.}
}
\value{
The plot of results for metabolite annotation of tandom mass spectrum.
}
\description{
Title annotaTandom_plot
}
\examples{
Parent_mass <- AnnotaData$Parent_mass
TandomData <- AnnotaData$TandomData
AnnotaParamTandom <- AnnotaTandom(Parent_mass, TandomData)
annotaDataTandom(AnnotaParamTandom)
annotaTandom_plot(AnnotaParamTandom, TandomData)
}
