% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step5_Enrichment.R
\name{Enrich_Plot}
\alias{Enrich_Plot}
\title{Title Enrich_Plot}
\usage{
Enrich_Plot(db, EnrichResultList)
}
\arguments{
\item{db}{The database of input data for enrichment analysis for other seven categories.}

\item{EnrichResultList}{The table of enrichment analysis for other seven categories.}
}
\value{
The plot of enrichment analysis for other seven categories.
}
\description{
Title Enrich_Plot
}
\examples{
sampleDatacas <- EnrichData$sampleDatacas
enrichDB <- EnrichData$enrichDB
EnrichParam <- KEGG_Enrich_PlotPanel(sampleDatacas, enrichDB = enrichDB, pvalcutoff = 0.05, IDtype = 2, cateIdx = 1)
EnrichResultList <- Enrichment(EnrichParam)
dbChoice <- enrichDB
Enrich_Plot(dbChoice,EnrichResultList)
}
