% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step5_Enrichment.R
\name{Enrichment}
\alias{Enrichment}
\title{Title}
\usage{
Enrichment(param)
}
\arguments{
\item{param}{The parameters for metabolite enrichment.}
}
\value{
The table of results for metabolite enrichment.
}
\description{
Title
}
\examples{
sampleDatakegg <- EnrichData$sampleDatakegg
EnrichParam <- KEGGEnrichPlotPanel(sampleDatakegg, enrichDB = "kegg", pvalcutoff = 0.05, IDtype = 1, cateIdx = 1)
Enrichment(EnrichParam)
}
