% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step5_Enrichment.R
\name{EnrichPlot}
\alias{EnrichPlot}
\title{Title}
\usage{
EnrichPlot(db, EnrichResultList)
}
\arguments{
\item{db}{The database of input for metabolite enrichment.}

\item{EnrichResultList}{The table of results for metabolite enrichment.}
}
\value{
The plot of results for metabolite enrichment.
}
\description{
Title
}
\examples{
sampleDatacas <- EnrichData$sampleDatacas
enrichDB <- EnrichData$enrichDB
EnrichParam <- KEGGEnrichPlotPanel(sampleDatacas, enrichDB = enrichDB, pvalcutoff = 0.05, IDtype = 2, cateIdx = 1)
EnrichResultList <- Enrichment(EnrichParam)
dbChoice <- enrichDB
EnrichPlot(dbChoice,EnrichResultList)
}
