% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step1_data_integrate.R
\name{Integrate_Data}
\alias{Integrate_Data}
\title{Title Integrate_Data}
\usage{
Integrate_Data(
  mutile_Group,
  RT_Tolerance_1 = 10,
  mz_Tolerance_1 = 0.1,
  RT_Tolerance_2 = 10,
  mz_Tolerance_2 = 0.1
)
}
\arguments{
\item{mutile_Group}{Multiple datasets from multiple groups for integrating as a comprehensive dataset.}

\item{RT_Tolerance_1}{Set the tolerance of retention time for data integration in the primary phase.}

\item{mz_Tolerance_1}{Set the tolerance of mass-to-charge ratio for data integration in the primary phase.}

\item{RT_Tolerance_2}{Set the tolerance of retention time for data integration in the secondary phase.}

\item{mz_Tolerance_2}{Set the tolerance of mass-to-charge ratio for data integration in the secondary phase.}
}
\value{
A matrix of a comprehensive dataset integrated form multiple groups is returned.
}
\description{
Title Integrate_Data
}
\examples{
Integrate_Data(mutile_Group, RT_Tolerance_1 = 10, mz_Tolerance_1 = 0.1, RT_Tolerance_2 = 10, mz_Tolerance_2 = 0.1)
}
