% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step5_Enrichment.R
\name{KEGG_Enrich_Plot}
\alias{KEGG_Enrich_Plot}
\title{Title KEGG_Enrich_Plot}
\usage{
KEGG_Enrich_Plot(EnrichResultList, cpdID, cpdFC)
}
\arguments{
\item{EnrichResultList}{The table of enrichment analysis for KEGG pathway.}

\item{cpdID}{The character of input data for enrichment analysis.}

\item{cpdFC}{The distance for the plot of enrichment analysis for KEGG pathway.}
}
\value{
The plot of enrichment analysis for KEGG pathway.
}
\description{
Title KEGG_Enrich_Plot
}
\examples{
sampleDatakegg <- EnrichData$sampleDatakegg
EnrichParam <- KEGG_Enrich_PlotPanel(sampleDatakegg, enrichDB = "kegg", pvalcutoff = 0.05, IDtype = 1, cateIdx = 1)
EnrichResultList <- Enrichment(EnrichParam)
EnrichFC <- seq(from=-2,to=2, length.out =24)
KEGG_Enrich_Plot(EnrichResultList=EnrichResultList,cpdID=sampleDatakegg,cpdFC=EnrichFC)
}
