% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step5_Enrichment.R
\name{KEGG_Enrich_PlotPanel}
\alias{KEGG_Enrich_PlotPanel}
\title{Title KEGG_Enrich_PlotPanel}
\usage{
KEGG_Enrich_PlotPanel(
  sampleData,
  enrichDB = "kegg",
  pvalcutoff = 0.05,
  IDtype = 1,
  cateIdx = 1
)
}
\arguments{
\item{sampleData}{The character of input data for enrichment analysis.}

\item{enrichDB}{The database for enrichment analysis. The database can be kegg, smpdb, cfam, foodb, biofunc, tcm, spectax or toxin.}

\item{pvalcutoff}{The cutoff of p value for enrichment analysis.}

\item{IDtype}{The number of type for names (1, 2, 3…) for enrichment analysis. If enrichDB is 'kegg', IDtype is 1 (KEGG ID), 2 (CAS ID), 3 (PubChem ID), or 4 (Compound Name). If enrichDB is 'smpdb', IDtype is 1 (HMDB ID), 2 (KEGG ID), 3 (CAS ID), 4 (PubChem ID), or 5 (Compound Name). If enrichDB is 'cfam', IDtype is 1 (CFAM ID), 2 (CAS ID), 3 (PubChem ID), 4 (Compound Name), or 5 (HMDB ID). If enrichDB is 'foodb', IDtype is 1 (FoodDB ID), 2 (CAS ID), 3 (PubChem ID), 4 (Compound Name) or 5 (HMDB ID). If enrichDB is 'biofunc', IDtype is 1 (HMDB ID), 2 (KEGG ID), 3 (CAS ID), 4 (PubChem ID) or 5 (Compound Name). If enrichDB is 'tcm', IDtype is 1 (LargeMetabo ID), 2 (PubChem ID), 3(Compound Name), or 4 (CAS ID). If enrichDB is 'spectax', IDtype is 1 (LargeMetabo ID), 2 (PubChem ID), 3(Compound Name), 4 (CAS ID) or 5 (HMDB ID). If enrichDB is 'toxin', IDtype is 1 (T3DB ID), 2 (PubChem ID), 3 (Compound Name), 4 (CAS ID) 5 (HMDB ID) or 6 (KEGG ID).}

\item{cateIdx}{The number (1, 2, 3…) of category for metabolite enrichment. If enrichDB is 'cfam', cateIdx is 1 (Class), 2 (Superfamily), or 3 (Family). If enrichDB is 'foodb', cateIdx is 1 (Food group), or 2 (Food subgroup). If enrichDB is 'biofunc', cateIdx is 1 (Blood), 2 (Urine), or 3 (Cerebrospinal). If enrichDB is 'tcm', cateIdx is 1 (TCM primary category), or 2 (TCM secondary category). If enrichDB is 'spectax', cateIdx is 1 (Superkingdom), 2 (Kingdom), 3(Phylum), 4 (Class), 5 (Order), 6 (Family), 7 (Genus), or 8 (Species).}
}
\value{
The parameters for enrichment analysis.
}
\description{
Title KEGG_Enrich_PlotPanel
}
\examples{
sampleDatakegg <- EnrichData$sampleDatakegg
KEGG_Enrich_PlotPanel(sampleDatakegg, enrichDB = "kegg", pvalcutoff = 0.05, IDtype = 1, cateIdx = 1)
}
