% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step5_Enrichment.R
\name{KEGGEnrichPlot}
\alias{KEGGEnrichPlot}
\title{Title KEGGEnrichPlot}
\usage{
KEGGEnrichPlot(EnrichResultList, cpdID, cpdFC)
}
\arguments{
\item{EnrichResultList}{The table of results for metabolite enrichment using KEGG database.}

\item{cpdID}{The charactor of input for metabolite enrichment.}

\item{cpdFC}{The distance in the metabolite enrichment plot using KEGG database.}
}
\value{
The plot of results for metabolite enrichment using KEGG database.
}
\description{
Title KEGGEnrichPlot
}
\examples{
sampleDatakegg <- EnrichData$sampleDatakegg
EnrichParam <- KEGGEnrichPlotPanel(sampleDatakegg, enrichDB = "kegg", pvalcutoff = 0.05, IDtype = 1, cateIdx = 1)
EnrichResultList <- Enrichment(EnrichParam)
EnrichFC <- seq(from=-2,to=2, length.out =24)
KEGGEnrichPlot(EnrichResultList=EnrichResultList,cpdID=sampleDatakegg,cpdFC=EnrichFC)
}
