% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step3_Marker_Identification.R
\name{Marker_Assess}
\alias{Marker_Assess}
\title{Title}
\usage{
Marker_Assess(finalData, finalLabel, method = "FC")
}
\arguments{
\item{finalData}{The matrix of dataset for marker identification.}

\item{finalLabel}{The label of dataset for marker identification.}

\item{method}{The method for marker identification. The method could be FC, PLS-DA, OPLS-DA, t-test, CHIS, CFS, ENTROPY, LMEB, RELIEF, RF, SAM, SVMRFE or WRST.}
}
\value{
The result of marker identification.

The ROC curve and AUC value in the SVM classification using markers.
}
\description{
Title
}
\examples{
finalData <- MarkerData$finalData
finalLabel <- MarkerData$finalLabel
Marker_Identify(finalData, finalLabel, method = "FC")
Title

finalData <- MarkerData$finalData
finalLabel <- MarkerData$finalLabel
Marker_Assess(finalData, finalLabel, method = "FC")

}
