% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step3_Marker_Identification.R
\name{Marker_Identify}
\alias{Marker_Identify}
\title{Title Marker_Identify}
\usage{
Marker_Identify(finalData, finalLabel, method = "FC")
}
\arguments{
\item{finalData}{The matrix of dataset for marker identification.}

\item{finalLabel}{The label of dataset for marker identification.}

\item{method}{The method for marker identification. The method can be FC, PLS-DA, OPLS-DA, t-test, CHIS, CFS, ENTROPY, LMEB, RELIEF, RF, SAM, SVMRFE or WRST.}
}
\value{
The result of marker identification for large-scale metabolomic data.
}
\description{
Title Marker_Identify
}
\examples{
finalData <- MarkerData$finalData
finalLabel <- MarkerData$finalLabel
Marker_Identify(finalData, finalLabel, method = "FC")
}
