% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step4_MetaboAnnotation.R
\name{Metabo_Annotation}
\alias{Metabo_Annotation}
\title{Title Metabo_Annotation}
\usage{
Metabo_Annotation(
  AnnotaData,
  masstole = 0.05,
  toleUnit = 1,
  annotaDB = "metlin",
  ionMode = "pos"
)
}
\arguments{
\item{AnnotaData}{The value of mass-to-charge ratio for MS1 spectra for metabolite annotation.}

\item{masstole}{The tolerance of m/z for metabolite annotation, and the unit is ppm or Da.}

\item{toleUnit}{The unit of m/z for metabolite annotation. The unit is 1 for Da and 2 for ppm.}

\item{annotaDB}{The database for metabolite annotation. The database can be metlin or hmdb.}

\item{ionMode}{The mode of m/z for metabolite annotation. The mode can be pos, neg or neu.}
}
\value{
The result of metabolite annotation for the adducts and charges for MS1 spectra.
}
\description{
Title Metabo_Annotation
}
\examples{
AnnotaMS <- AnnotaData$AnnotaMS
MetaboAResult <- Metabo_Annotation(AnnotaMS)
MetaboAResult$`M+H-2H2O`


}
