% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step1_data_integrate.R
\name{Removal_Batch}
\alias{Removal_Batch}
\title{Title Removal_Batch}
\usage{
Removal_Batch(mutile_align, n = 3, algorithm = "BMC/PAMR")
}
\arguments{
\item{mutile_align}{The comprehensive dataset integrated from multiple groups.}

\item{n}{The number of multiple groups.}

\item{algorithm}{The algorithm for removing batch effects in multiple groups. The algorithm could be BMC/PAMR, ComBat/EB, DWD, GlobalNorm, XPN or None.}

\item{loga}{TRUE indicates log transformation will be used, FALSE indicates log transformation will not be used.}
}
\value{
A matrix of a dataset after removing batch effects for the integrated dataset.
}
\description{
Title Removal_Batch
}
\examples{
Removal_Batch(mutile_align, n = 3, algorithm = "BMC/PAMR", loga= TRUE)
}
