% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step2_Sample_Separation.R
\name{Sample_Separation}
\alias{Sample_Separation}
\title{Title Sample_Separation}
\usage{
Sample_Separation(finalData, finalLabel, clusters = 2, method = "HCA")
}
\arguments{
\item{finalData}{The matrix of dataset for sample separation.}

\item{finalLabel}{The label of dataset for sample separation.}

\item{clusters}{The number of cluster for sample separation.}

\item{method}{The method for sample separation. The method could be HCA, KMC, PCA and SOM.}
}
\value{
The plot of sample separation.
}
\description{
Title Sample_Separation
}
\examples{
finalData <- MarkerData$finalData
finalLabel <- MarkerData$finalLabel
Sample_Separation(finalData, finalLabel, clusters=2, method = "HCA")
}
