% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step5_Enrichment.R
\name{Enrichment}
\alias{Enrichment}
\title{Title Enrichment}
\usage{
Enrichment(param)
}
\arguments{
\item{param}{The parameters for enrichment analysis.}
}
\value{
The table of results for enrichment analysis.
}
\description{
Title Enrichment
}
\examples{
sampleDatakegg <- EnrichData$sampleDatakegg
EnrichParam <- KEGG_Enrich_PlotPanel(sampleDatakegg, enrichDB = "kegg", pvalcutoff = 0.05, IDtype = 1, cateIdx = 1)
EnrichResultList <- Enrichment(EnrichParam)
EnrichResultList$Table.Result
}
