% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step1_data_integrate.R
\name{Integrate_Data}
\alias{Integrate_Data}
\title{Title Integrate_Data}
\usage{
Integrate_Data(
  MutileGroup,
  RTTolerance1 = 10,
  mzTolerance1 = 0.1,
  RTTolerance2 = 10,
  mzTolerance2 = 0.1
)
}
\arguments{
\item{MutileGroup}{Data list including multiple datasets from multiple experiments and are integrated as a comprehensive dataset.}

\item{RTTolerance1}{Set the tolerance of retention time in the first step of data integration.}

\item{mzTolerance1}{Set the tolerance of mass-to-charge ratio in the first step of data integration.}

\item{RTTolerance2}{Set the tolerance of retention time in the second step of data integration.}

\item{mzTolerance2}{Set the tolerance of mass-to-charge ratio in the second step of integration.}
}
\value{
A matrix of an integrated dataset form multiple analytical experiments is returned.
}
\description{
Title Integrate_Data
}
\examples{
Integrate_Data(MutileGroup, RTTolerance1 = 10, mzTolerance1 = 0.1, RTTolerance2 = 10, mzTolerance2 = 0.1)
}
