% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step3_Marker_Identification.R
\name{Marker_Assess}
\alias{Marker_Assess}
\title{Title Marker_Assess}
\usage{
Marker_Assess(finalData, finalLabel, method = "PLS-DA")
}
\arguments{
\item{finalData}{The matrix of dataset for performance assessment of marker identification.}

\item{finalLabel}{The label of dataset for performance assessment of marker identification.}

\item{method}{The method used for identifying markers for performance assessment. The method can be FC, PLS-DA, OPLS-DA, t-test, CHIS, CFS, ENTROPY, LMEB, RELIEF, RF, SAM, SVMRFE or WRST.}
}
\value{
The ROC curve and AUC value in the SVM classification model using identified markers.
}
\description{
Title Marker_Assess
}
\examples{
finalData <- MarkerData$finalData
finalLabel <- MarkerData$finalLabel
Marker_Assess(finalData, finalLabel, method = "PLS-DA")

}
