% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step1_data_integrate.R
\name{Removal_Batch}
\alias{Removal_Batch}
\title{Title Removal_Batch}
\usage{
Removal_Batch(MutileAlign, n = 3, algorithm = "BMC/PAMR")
}
\arguments{
\item{MutileAlign}{The integrated dataset from multiple analytical experiments.}

\item{n}{The number of multiple analytical experiments.}

\item{algorithm}{The algorithm for removing batch effects in multiple analytical experiments. The algorithm can be BMC/PAMR, ComBat/EB, GlobalNorm or None.}
}
\value{
A matrix of a dataset after removing batch effects for the integrated dataset.
}
\description{
Title Removal_Batch
}
\examples{
Removal_Batch(MutileAlign, n = 3, algorithm = "BMC/PAMR")
}
